Renamed Language to Settings
In the next commit I will add new relevant settings to the class and make the class more modular.

I also got rid of the reportWebVitals because I don't think we'll use it, and made a scroll bar always appear because it appearing only sometimes was kind of ugly
