+++ src/English.js
@@ -1,5 +1,5 @@
 import { Component } from 'react';
-import {language, DEFAULT_ENGLISH_VISIBILITY} from './Language';
+import {settings, DEFAULT_ENGLISH_VISIBILITY} from './Settings';
     
 class English extends Component {
     constructor(props){ 
@@ -9,10 +9,10 @@ class English extends Component {
     }
 
     componentDidMount () {
-        this.removeListener = language.addListener((state) => {
+        this.removeListener = settings.addListener((state) => {
             this.setState({englishVisible: state.englishVisible});
         });
-        this.setState({englishVisible: language.getState().englishVisible});
+        this.setState({englishVisible: settings.getState().englishVisible});
     }
     componentWillUnmount() {
         this.removeListener();
