+++ src/global.css
@@ -1,6 +1,7 @@
 body {
   background-color: rgb(254, 254, 254);
   color: black;
+  overflow-y: scroll;
 }
 
 #Nav-spacing {
